Bug repair

+++ app/src/main/java/lab/android/bartosz/ssms/SensorDataDbHelper.java
@@ -165,7 +165,7 @@ public class SensorDataDbHelper extends SQLiteOpenHelper {
 
     }
 
-    public SensorData getByID(int id)
+    public SensorData getByID(long id)
     {
 
         SensorData sensorData = new SensorData();
